requirements.txt
colorama==0.4.6
Flask==3.0.2
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
flask-sitemapper==1.7.0
Jinja2==3.1.3
persiantools==3.0.1
PyMySQL==1.1.0
requests==2.30.0
Werkzeug==3.0.1